+++ packages/lenses/src/index.ts
@@ -1,4 +1,5 @@
 export * from "./availability";
 export * from "./render-markup";
+export * from "./safe-url";
 export * from "./source-symbols";
 export * from "./user-data";